career_path/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
sqlparse==0.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
numpy==2.2.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
threadpoolctl==3.5.0
urllib3==2.3.0
pandas==2.0.3
surprise==0.1